requirements.txt
certifi==2019.6.16
Django==2.2.1
django-autofixture==0.12.1
django-infinite-scroll-pagination==1.0.0
mkl-fft==1.0.12
mkl-random==1.0.2
numpy==1.16.4
olefile==0.46
pandas==0.24.2
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
xlrd==1.2.0
